+++ .vscode/launch.json
@@ -23,7 +23,7 @@
         "package",
         "file://${workspaceFolder}/fixtures/schema/draft-2020-12/schema",
         "--package-directory",
-        ".package/draft-2020-12",
+        ".package/schema/draft-2020-12",
         "--package-name",
         "draft-2020-12",
         "--package-version",
@@ -45,7 +45,7 @@
         "package",
         "file://${workspaceFolder}/fixtures/schema/draft-2019-09/schema",
         "--package-directory",
-        ".package/draft-2019-09",
+        ".package/schema/draft-2019-09",
         "--package-name",
         "draft-2019-09",
         "--package-version",
@@ -67,7 +67,7 @@
         "package",
         "file://${workspaceFolder}/fixtures/schema/draft-07/schema",
         "--package-directory",
-        ".package/draft-07",
+        ".package/schema/draft-07",
         "--package-name",
         "draft-07",
         "--package-version",
@@ -89,7 +89,7 @@
         "package",
         "file://${workspaceFolder}/fixtures/schema/draft-06/schema",
         "--package-directory",
-        ".package/draft-06",
+        ".package/schema/draft-06",
         "--package-name",
         "draft-06",
         "--package-version",
@@ -111,7 +111,7 @@
         "package",
         "file://${workspaceFolder}/fixtures/schema/draft-04/schema",
         "--package-directory",
-        ".package/draft-04",
+        ".package/schema/draft-04",
         "--package-name",
         "draft-04",
         "--package-version",
