+++ fixtures/testing/schema/draft-07/string-or-boolean.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$schema": "https://json-schema.org/draft-07/schema#",
   "type": [
     "string",
     "boolean"
