+++ fixtures/testing/schema/draft-2020-12/array-with-ref.json
@@ -1,6 +1,5 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://JsonSchema42.org/array-with-ref",
   "description": "So this is an array!",
   "type": [
     "array"
