+++ fixtures/testing/schema/draft-2020-12/simple-object.json
@@ -1,6 +1,5 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://JsonSchema42.org/simple-object",
   "description": "A very simple object",
   "type": [
     "object"
