+++ src/schema/draft-2020-12/strategy.ts
@@ -54,7 +54,10 @@ export class SchemaStrategy extends SchemaStrategyBase<Schema> {
             return nodeUrl;
         }
 
-        nodeUrl = new URL(`#${nodePointer}`, nodeRootUrl);
+        nodeUrl = new URL(
+            nodePointer === "" ? "" : `#${nodePointer}`,
+            nodeRootUrl,
+        );
         return nodeUrl;
     }
